backend/requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
psycopg2-binary==2.9.9
reportlab==4.0.7
python-dotenv==1.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
SQLAlchemy==2.0.23
Jinja2==3.1.2
PyPDF2==3.0.1
